% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.define}
\alias{as.define}
\alias{as.define.spec}
\title{Coerce to Define}
\usage{
as.define(x, ...)

\method{as.define}{spec}(x, sep = " = ", collapse = "; ",
  escape = character(0), ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}

\item{sep}{separates codes from respective decodes where given}

\item{collapse}{separates code/decode pairs where given}

\item{escape}{values to escape for proper latex formatting}
}
\description{
Coerces to class 'define'.  Generic, with method for 'spec'.

Coerces to class 'define' from class 'spec'. Extra arguments ignored.
}
\section{Methods (by class)}{
\itemize{
\item \code{spec}: method for spec
}}

\seealso{
\code{\link[spec]{specification}}

\code{\link[spec]{as.spec}}
}
