% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{as.labeled.character}
\alias{as.labeled.character}
\title{Coerce character to class labeled.}
\usage{
\method{as.labeled}{character}(x, spec, as.is = TRUE, na.strings = c("",
  "."), rename = function(x, ...) x, ...)
}
\arguments{
\item{x}{length-one filename for csv-formatted data file}

\item{spec}{length-one file name for spec}

\item{as.is}{passed to read.csv}

\item{na.strings}{passed to read.csv}

\item{rename}{a function with arguments x, ... to pre-process column names}

\item{...}{passed to as.labeled.dat}
}
\description{
Coerces character to a labeled data.frame.
}
\seealso{
\code{\link{as.labeled.data.frame}}
}

