% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{as.labeled}
\alias{as.labeled}
\alias{as.labeled.character}
\title{Coerce to class labeled.}
\usage{
as.labeled(x, ...)

\method{as.labeled}{character}(
  x,
  spec,
  as.is = TRUE,
  na.strings = c("", "."),
  rename = function(x, ...) x,
  ...
)
}
\arguments{
\item{x}{length-one filename for csv-formatted data file}

\item{...}{passed to as.labeled.dat}

\item{spec}{length-one file name for spec}

\item{as.is}{passed to read.csv}

\item{na.strings}{passed to read.csv}

\item{rename}{a function with arguments x, ... to pre-process column names}
}
\description{
Coerces to class labeled.

Coerces character to a labeled data.frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: data.frame method for as.labeled
}}

\seealso{
\code{\link{as.labeled.data.frame}}
}
