% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.pdf.define}
\alias{as.pdf.define}
\title{Coerce to PDF from Define}
\usage{
\method{as.pdf}{define}(x, stem, ...)
}
\arguments{
\item{x}{define object}

\item{stem}{passed to \code{\link[latexpdf]{as.pdf}}}

\item{...}{passed to \code{\link[latexpdf]{as.pdf}} and \code{\link[latexpdf]{as.document}}}
}
\description{
Coerces to PDF from class 'define'.
}
