% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{as.submission}
\alias{as.submission}
\alias{as.submission.character}
\title{Create a submission.}
\usage{
as.submission(x, ...)

\method{as.submission}{character}(x, tag = names(x),
  description = basename(x), dir = ".", subdir = NULL, ...)
}
\arguments{
\item{x}{filenames: xpt, csv, spec, txt, other}

\item{...}{passed along to handlers}

\item{tag}{short names for each element of x}

\item{description}{informative multi-word label for each element of x}

\item{dir}{parent directory for placement of submission artifacts}

\item{subdir}{optional subdirectories relative to dir for each submission artifact}
}
\value{
a list of artifacts each having attributes: x, tag, des, file, spec
}
\description{
Returns a list of vetted artifacts (spec or char corresponding to x)
that represent a submission object for further processing.

Converts a character vector of file names to class submission.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}

