% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define.R
\name{as.submission.submission}
\alias{as.submission.submission}
\title{Coerce a submission to class submission.}
\usage{
\method{as.submission}{submission}(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed along}
}
\description{
Coerces a submission to class submission.
}

