% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.tabular.define}
\alias{as.tabular.define}
\title{Coerce to Tabular from Define}
\usage{
\method{as.tabular}{define}(x, caption = "", grid = TRUE, rules = 1,
  colwidth = c("1in", "1in", "0.5in", "1.5in", "1.5in"),
  tabularEnvironment = "longtable", walls = 1, tabnum = FALSE,
  pretable = if (is.null(caption)) "" else paste(if (tabnum) "\\\\caption{"
  else "\\\\caption*{", caption,
  "}\\\\\\\\"), prepos = 1, headerBold = TRUE, ...)
}
\arguments{
\item{x}{define object}

\item{caption}{caption for definitions table}

\item{grid}{passed to \code{\link[latexpdf]{as.tabular.data.frame}}}

\item{rules}{passed to \code{\link[latexpdf]{as.tabular.data.frame}}}

\item{colwidth}{passed to \code{\link[latexpdf]{as.tabular.data.frame}}}

\item{tabularEnvironment}{passed to \code{\link[latexpdf]{as.tabular.data.frame}}}

\item{walls}{passed to \code{\link[latexpdf]{as.tabular.data.frame}}}

\item{tabnum}{whether to number the table}

\item{pretable}{material to include before table, typically a caption}

\item{prepos}{after which line number should pretable be inserted?}

\item{headerBold}{whether to use a bold header}

\item{...}{passed to \code{\link[latexpdf]{as.tabular}}}
}
\description{
Coerces to class 'tabular' from class 'define'.
}
