% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define.R
\name{as.xport}
\alias{as.xport}
\alias{as.xport.labeled}
\title{Coerce to class xport.}
\usage{
as.xport(x, ...)

\method{as.xport}{labeled}(x, name, file, autogen.formats = FALSE, ...)
}
\arguments{
\item{x}{a labeled data.frame}

\item{...}{passed to write.xport}

\item{name}{a name for the data.frame}

\item{file}{where to write the data.frame}

\item{autogen.formats}{passed to write.xport}
}
\description{
Coerces to class xport.

Coerces labeled to xport.
}
\section{Methods (by class)}{
\itemize{
\item \code{labeled}: 
}}

