% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define.character}
\alias{define.character}
\title{Define a set of files per FDA guidance.
 
 tag is taken by default as the names of x, but may be supplied explicitly.
 The following should have the same length as x: tag, description, subdir.
 subdir may also be length one. Other arguments have length one.
 
 The function iterates across the elements of x to create a submission
 object, a side effect of which is to copy (conditionally, transformed) each 
 corresponding file to (subdir of) dir.  The submission object is then
 converted to a pdf, written directly to dir as <stem>.pdf.
 
 Arguments short, protocol, sponsor, program, author, and date are length-one
 character that define attributes of the pdf title page. Alternatively, 
 they may be specified by names that reflect position rather than semantics:
 lhead1, lhead2, rhead1, rhead2, lfoot, rfoot (respectively).
 
 @seealso \url{http://tinyurl.com/fda-pdf-spec-4-0}
 @seealso \url{http://tinyurl.com/fda-study-data-spec-2-0}
 
 @param x paths to existing files to be documented; possibly a named character vector
 @param stem the base of the file name for the pdf to be created
 @param tag short object names for each element of x; appears in pdf menu, and as table name in XPT file
 @param description informative labels for each element of x
 @param title a title to appear in the pdf
 @param short short title a.k.a. lhead1 (upper left pdf header)
 @param protocol relevant protocols a.k.a. lhead2 (lower left pdf header)
 @param sponsor study sponsor a.k.a. rhead1 (upper right pdf header)
 @param program drug development program a.k.a. rhead2 (lower right pdf header)
 @param author document author a.k.a. lfoot left pdf footer, italicized)
 @param date date format string a.k.a. rfoot (right footer) today's date by default
 @param logo file path for logo to include on cover page
 @param logoscale size adjustment for logo
 @param dir path to directory in which to place pdf and copied (transformed) files
 @param subdir path to subdirectories to which to copy each (transformed) file represented by x; use NULL to suppress archiving
 @param clear should dir be deleted if it exists?
 @param ... passed to as.submission and as.pdf
 
 @return invisible result of as.pdf.  Used for side effects.
 @export
 @describeIn define
 @examples 
 
code <- "write.csv(x=Theoph,file='theoph.csv',row.names=FALSE,quote=FALSE)"
writeLines(code,'theoph.R')
eval(parse(text=code))
write.spec(specification(Theoph),'theoph.spec')
define(c(theoph='theoph.csv'),stem='minimal')
as.pdf('{\\huge \\em Pharma, Inc.}',wide=50, long = 8,stem='logo')
define(
  x=c(
    theodat='theoph.csv',
    theoprg='theoph.R'
   ),
  subdir=c(
    'm5/datasets/analysis/datasets',
    'm5/datasets/analysis/programs'
  ),
  description=c(
    'Theophylline PK Dataset',
    'Theophylline PK Script'
  ),
  title = 'Pharmacokinetics of Theophylline',
  short = 'Theophylline PK',
  protocol = 'Protocol tpk-001',
  sponsor = 'Pharma, Inc.',
  program = 'Theophylline',
  author = 'define package for R',
  logo = 'logo.pdf',
  logoscale = 2,
  clear=FALSE
)}
\usage{
\method{define}{character}(x, stem = "define", tag = names(x),
  description = basename(x), title = dirname(x[[1]]), short = title,
  protocol = "~", sponsor = "~", program = "~", author = "~",
  date = "\\\\mydate \\\\today", logo = NULL, logoscale = 1,
  dir = "./define", subdir = ".", clear = TRUE, ...)
}
\description{
Define a set of files per FDA guidance.
 
 tag is taken by default as the names of x, but may be supplied explicitly.
 The following should have the same length as x: tag, description, subdir.
 subdir may also be length one. Other arguments have length one.
 
 The function iterates across the elements of x to create a submission
 object, a side effect of which is to copy (conditionally, transformed) each 
 corresponding file to (subdir of) dir.  The submission object is then
 converted to a pdf, written directly to dir as <stem>.pdf.
 
 Arguments short, protocol, sponsor, program, author, and date are length-one
 character that define attributes of the pdf title page. Alternatively, 
 they may be specified by names that reflect position rather than semantics:
 lhead1, lhead2, rhead1, rhead2, lfoot, rfoot (respectively).
 
 @seealso \url{http://tinyurl.com/fda-pdf-spec-4-0}
 @seealso \url{http://tinyurl.com/fda-study-data-spec-2-0}
 
 @param x paths to existing files to be documented; possibly a named character vector
 @param stem the base of the file name for the pdf to be created
 @param tag short object names for each element of x; appears in pdf menu, and as table name in XPT file
 @param description informative labels for each element of x
 @param title a title to appear in the pdf
 @param short short title a.k.a. lhead1 (upper left pdf header)
 @param protocol relevant protocols a.k.a. lhead2 (lower left pdf header)
 @param sponsor study sponsor a.k.a. rhead1 (upper right pdf header)
 @param program drug development program a.k.a. rhead2 (lower right pdf header)
 @param author document author a.k.a. lfoot left pdf footer, italicized)
 @param date date format string a.k.a. rfoot (right footer) today's date by default
 @param logo file path for logo to include on cover page
 @param logoscale size adjustment for logo
 @param dir path to directory in which to place pdf and copied (transformed) files
 @param subdir path to subdirectories to which to copy each (transformed) file represented by x; use NULL to suppress archiving
 @param clear should dir be deleted if it exists?
 @param ... passed to as.submission and as.pdf
 
 @return invisible result of as.pdf.  Used for side effects.
 @export
 @describeIn define
 @examples 
 
code <- "write.csv(x=Theoph,file='theoph.csv',row.names=FALSE,quote=FALSE)"
writeLines(code,'theoph.R')
eval(parse(text=code))
write.spec(specification(Theoph),'theoph.spec')
define(c(theoph='theoph.csv'),stem='minimal')
as.pdf('{\\huge \\em Pharma, Inc.}',wide=50, long = 8,stem='logo')
define(
  x=c(
    theodat='theoph.csv',
    theoprg='theoph.R'
   ),
  subdir=c(
    'm5/datasets/analysis/datasets',
    'm5/datasets/analysis/programs'
  ),
  description=c(
    'Theophylline PK Dataset',
    'Theophylline PK Script'
  ),
  title = 'Pharmacokinetics of Theophylline',
  short = 'Theophylline PK',
  protocol = 'Protocol tpk-001',
  sponsor = 'Pharma, Inc.',
  program = 'Theophylline',
  author = 'define package for R',
  logo = 'logo.pdf',
  logoscale = 2,
  clear=FALSE
)
}

