% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define}
\alias{define}
\title{Define objects per FDA guidance.}
\usage{
define(x, ...)
}
\description{
Defines (documents) a set of files in a manner intended to comply with 
 FDA guidance on submission of study data and related documentation.
 In particular, files in csv format are converted to SAS Transport (xpt),
 extensions for other files (presumably ASCII) are coerced to txt, 
 files are copied to a directory tree, and 
 define.pdf is created at the top level to describe the files in more detail.
}

