% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{makesasnames}
\alias{makesasnames}
\title{Make names for a dataset that are unique and follow SAS naming conventions.}
\usage{
makesasnames(names, nchar = 8, maxPasses = 10, quiet = FALSE)
}
\arguments{
\item{names}{existing column names}

\item{nchar}{limit on number of characters}

\item{maxPasses}{limit on number of reconciliation attempts}

\item{quiet}{should messages be suppressed?}
}
\value{
character
}
\description{
Makes names for a dataset that are unique and follow SAS naming conventions.
Modeled after SASxport::makeSasNames, but handling special cases.
}

