% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/define.R
\name{as.labeled}
\alias{as.labeled}
\alias{as.labeled.character}
\title{Coerce to class labeled.}
\usage{
as.labeled(x, ...)

\method{as.labeled}{character}(x, spec, ...)
}
\arguments{
\item{x}{length-one filename for csv-formatted data file}

\item{...}{passed to as.labeled.dat}

\item{spec}{length-one file name for spec}
}
\description{
Coerces to class labeled.

Coerces character to a labeled data.frame.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: 
}}
\seealso{
\code{\link{as.labeled.data.frame}}
}

