% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\docType{package}
\name{define-package}
\alias{define-package}
\title{Create FDA-style dataset and program definitions.}
\description{
\pkg{define} helps you create 'define.pdf' and associated file tree
for FDA-style submission of analysis datasets, etc.
It converts csv files to SAS Transport V. 5 'xpt' format,
using metadata encoded in a specification file. It enforces
the 'txt' extension for other (presumably ASCII) files.
}
\details{
The only function you're likely to need from \pkg{define} is
\code{\link{define}}. You may want to learn more about
metadata encoding from \pkg{encode} and more about 
specification files from \pkg{spec}.
}
\references{
\href{https://www.fda.gov/media/83880/download}{FDA Study Data Specifications}

\href{http://tinyurl.com/onocmc5}{FDA PDF Specifications}

\href{https://support.sas.com/techsup/technote/ts140.pdf}{SAS Transport Format Specification}
}
\author{
Tim Bergsma, \email{bergsmat@gmail.com}
}
